# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sable Operator API
  version: 1.2.0
  description: |
    Loopback-only operator API for the Sable C2 server. Bound to
    `127.0.0.1:8443` by default. All endpoints except `POST /api/auth/login`
    require a Bearer JWT obtained from `/api/auth/login`.

    **Error responses are plain text** (`text/plain; charset=utf-8`), not JSON.
    Each error response below documents the canonical message body emitted by
    the server.

    The TLS certificate is self-signed and pinned by SHA-256 fingerprint when
    operators connect through `sablectl`. Browsers will need to accept the
    certificate manually.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: https://127.0.0.1:8443
    description: Operator API on the Sable server (loopback)
security:
  - bearerAuth: []
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Exchange operator password for a JWT
      description: |
        Returns a signed HS256 JWT valid for 1 hour. Rate-limited to 5
        attempts per 60 seconds per remote IP; rate-limited responses
        include a `Retry-After` header.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              password: correct-horse-battery-staple
      responses:
        '200':
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/RequestEntityTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429':
          description: Too many login attempts from this IP.
          headers:
            Retry-After:
              schema: { type: integer }
              description: Seconds until another attempt is allowed.
          content:
            text/plain:
              schema: { type: string }
              example: "too many requests\n"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 JWT issued by `/api/auth/login`. Issuer `sable-operator`,
        audience `sable-web`, subject `operator`, 1-hour TTL.
...
  schemas:
    LoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: Operator password. Request body capped at 4096 bytes.
          example: correct-horse-battery-staple
    LoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Bearer JWT (HS256). Valid for 1 hour.
...
  responses:
    BadRequest:
      description: Malformed request, validation failure, or invalid task type.
      content:
        text/plain:
          schema: { type: string }
          example: "bad request\n"
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        text/plain:
          schema: { type: string }
          example: "unauthorized\n"
...
    MethodNotAllowed:
      description: Method not supported on this path.
      content:
        text/plain:
          schema: { type: string }
          example: "method not allowed\n"
    RequestEntityTooLarge:
      description: Request body exceeded the per-endpoint byte cap.
      content:
        text/plain:
          schema: { type: string }
          example: "request body too large\n"
    UnsupportedMediaType:
      description: Body Content-Type is not `application/json`.
      content:
        text/plain:
          schema: { type: string }
          example: "content type must be application/json\n"